+++ .env
@@ -1,6 +1,7 @@
 NEXT_PUBLIC_CHROME_EXT_ID=fmblappgoiilbgafhjklehhfifbdocee
 NEXT_PUBLIC_COINGECKO_API_URL=https://api.coingecko.com/api/v3
 NEXT_PUBLIC_CHROME_EXT_INSTALL_URL=https://chrome.google.com/webstore/detail/forbole-x/fmblappgoiilbgafhjklehhfifbdocee
+NEXT_PUBLIC_IPFS_URL=https://ipfs-staging.desmos.network
 MAILGUN_HOST=smtp.mailgun.org
 MAILGUN_PORT=2525
 MAILGUN_USER=secret
